+++ CHANGELOG.md
@@ -10,6 +10,14 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.12.2] - 2024-09-30
+
+### Fixed
+
+- Fixed error in `Marshaler.MarshalAs()` when passed a media-type that is
+  unsupported because it does not have a `type` parameter. The implementation
+  now correctly returns `false` instead.
+
 ## [0.12.1] - 2024-09-29
 
 ### Added
@@ -104,6 +112,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.11.1]: https://github.com/dogmatiq/enginekit/releases/v0.11.1
 [0.12.0]: https://github.com/dogmatiq/enginekit/releases/v0.12.0
 [0.12.1]: https://github.com/dogmatiq/enginekit/releases/v0.12.1
+[0.12.2]: https://github.com/dogmatiq/enginekit/releases/v0.12.2
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD